fix controlled component for texfield
